fix: The special exam timer synced with backend to show accurate count down timer (#45)
This is a revert of a previous revert. This reverts commit e6aa31c100102235d04bd8d5507fc924a150719a. This PR would use the proper sermantic-release commit message.

The special exam timer shown to learners should be updated with the exam attempt remaining time available for each sync interval. This is to ensure our learners don't see a timer which is running slower than the actual time.

Co-authored-by: Simon Chen <schen@edx-c02fw0guml85.lan>