fix: prevent exam errors from displaying off exam page
Since we make an exam attempt call on every sequence to show the exam
bar if people have navigated away from it, we run the risk of 403 or
other transient problem in the exam call on every page. That then
shows an error which makes no sense when the user is not taking an
exam.

Since the errors are still critical when you are working on the exam
and if the call fails we don't know if you have a current exam, we've
decided to suppress these errors if you are not currently looking at
the exam itself. For most pages this prevents spurious errors. If you
are in an exam and navigate away from it and then have an error that
will also be suppressed, but you will recover from transient errors on
the next page and get the exam bar back.

MST-905